I made my first project website with a parallax effect using HTML and CSS
Some important points I learn while making this website are.
📌 Use of perspective property,
📌 Transform style property.
📌 Different types of positioning elements.
📌 Z index
📌 Background-attachment property


The Parallax effect is based on the technique in which we put images or any content at different distances from the screen (ie in-axis, inside the screen). Through this, we will see the difference in speed of different (images or content) while scrolling down the farther objects.