I created this Portfolio Website using html and css.
Excited to announce the completion of my latest project, a portfolio website built using HTML and CSS! It was a rewarding experience working with these technologies and I am proud of the final product. 

This project was not just about creating a portfolio, but also about learning and experimenting with new techniques.
I had the opportunity to work with some amazing CSS properties such as
 👉 keyframes,
 👉 @media,
 👉 pseudo-elements,
 👉 Transition 

These properties have made my portfolio project responsive. 
I am particularly excited about the animations that have been implemented, which have added a dynamic touch to the website